app/main.py
from fastapi import FastAPI, UploadFile, File, Form
from fastapi.responses import HTMLResponse
from fastapi.staticfiles import StaticFiles
import numpy as np
from pyod.models.auto_encoder import AutoEncoder
import tensorflow as tf
from transformers import BertTokenizer, TFBertModel
import google.cloud.storage
import google.cloud.bigquery

app = FastAPI()

# Mount static folder for frontend
app.mount("/static", StaticFiles(directory="static"), name="static")

# Placeholder for feature extraction models
image_model = tf.keras.applications.ResNet50(weights="imagenet", include_top=False)
tokenizer = BertTokenizer.from_pretrained("bert-base-uncased")
text_model = TFBertModel.from_pretrained("bert-base-uncased")

# Placeholder for anomaly detection model (tune later)
anomaly_model = AutoEncoder(hidden_neurons=[128, 64, 64, 128])

# GCP clients (configure credentials in Cloud Run environment)
storage_client = google.cloud.storage.Client()
bq_client = google.cloud.bigquery.Client()

@app.get("/", response_class=HTMLResponse)
async def serve_frontend():
    with open("static/index.html") as f:
        return HTMLResponse(content=f.read())

@app.post("/upload")
async def upload_data(
    image: UploadFile = File(...),
    text: str = Form(...),
    label: str = Form(default="unknown")
):
    # Step 1: Store file in Cloud Storage
    bucket = storage_client.bucket("your-bucket-name")  # Replace with your bucket
    blob = bucket.blob(f"uploads/{image.filename}")
    blob.upload_from_file(image.file)

    # Step 2: Feature extraction (placeholders, tune later)
    # Image processing
    img = tf.image.decode_image(await image.read(), channels=3)
    img = tf.image.resize(img, [224, 224])
    img_features = image_model.predict(tf.expand_dims(img, 0)).flatten()

    # Text processing
    inputs = tokenizer(text, return_tensors="tf", padding=True, truncation=True)
    text_features = text_model(inputs)[0][:, 0, :].numpy().flatten()

    # Combine features
    combined_features = np.concatenate([img_features, text_features])

    # Step 3: Anomaly detection (placeholder, train later)
    score = anomaly_model.decision_function(combined_features.reshape(1, -1))[0]

    # Step 4: Store in BigQuery (table creation manual)
    row = {"image_path": f"gs://your-bucket-name/uploads/{image.filename}", "text": text, "score": float(score), "label": label}
    bq_client.insert_rows_json("your_project.your_dataset.your_table", [row])  # Replace with your table

    # Step 5: Return result
    return {"score": score, "is_anomaly": score > 1.0}  # Threshold TBD

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)