app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pyod==2.0.2
tensorflow==2.17.0
transformers==4.44.2
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
numpy==1.26.4